refactor(serializers): swap TypeSerializer and ModelSerializer fields
Rearranged the serializers to correctly match models with their intended fields. The TypeSerializer now includes 'brands' and uses the Type model, while the ModelSerializer associates with the Model model and its respective fields. This improves serialization alignment with data models and enhances clarity of relationships between types, brands, and models.